Rename debugger.ts to debug_manager.ts
